# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vinmo API",
        "description": "Documentação da API do Vinmo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/images/{image}": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Exibe uma imagem",
                "operationId": "bb3175b3e40a396bcc3e52c5f7581065",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "ID da imagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stream da imagem"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Images"
                ],
                "summary": "Remove uma imagem",
                "operationId": "25687ab67804accc883dc67eeec16d21",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "ID da imagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagem removida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/images/{image}/download": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Baixa a imagem original",
                "operationId": "a3ceaedf365f1f2b46ebf66f56b2c619",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "ID da imagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Download da imagem"
                    }
                }
            }
        },
        "/api/images/{image}/metadata": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Recupera os metadados da imagem",
                "operationId": "27d2168a6a62464feb894a887e72f209",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "ID da imagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadados recuperados"
                    }
                }
            }
        },
        "/api/images/{image}/clients": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Retorna a lista de clientes identificados em uma imagem",
                "operationId": "06024e37ab06d28d42e807c9a525575f",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "ID da imagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clientes recuperados"
                    }
                }
            }
        },
        "/api/images/{image}/clients/{client}/crop": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Obtém o recorte do rosto do cliente na imagem",
                "operationId": "a996404dd897ef0561cee457b4871e74",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "ID da imagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recorte obtido"
                    }
                }
            }
        },
...
        "/api/locations/countries/{country_cca2}/states": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Obtém a lista de estados de um país",
                "operationId": "6dce5fd76bb73ac391ff8b0e557327dd",
                "parameters": [
                    {
                        "name": "country_cca2",
                        "in": "path",
                        "description": "Código CCA2 do país (ex: BR)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de estados obtida"
                    },
                    "404": {
                        "description": "País não encontrado"
                    }
                }
            }
        },
        "/api/locations/countries/{country_cca2}/states/{state_code}/cities": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Obtém a lista de cidades de um estado",
                "operationId": "0d75e8710e086c19923c93f1e23d3833",
                "parameters": [
                    {
                        "name": "country_cca2",
                        "in": "path",
                        "description": "Código CCA2 do país (ex: BR)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state_code",
                        "in": "path",
                        "description": "Código do estado (ex: SP)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de cidades obtida"
                    },
                    "404": {
                        "description": "País ou estado não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}